requirements.txt
Django==5.0.3
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.2.1
requests==2.32.5
Pillow==10.3.0
django-ckeditor==6.7.3
